Restructuring project with new folders
