+++ client/App.jsx
@@ -1,15 +1,20 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { NavBar } from "./src/Components/Navbar";
-import Produts from "./src/Components/Produts/Produts";
+import ProductCard from "./src/Components/ProductCard/ProductCard";
+import LandingPage from './src/views/LandingPage/LandingPage';
+import Home from './src/views/Home/Home';
+import Detail from './src/views/Detail/Detail'
 
 const App = () => {
   return (
     <>
-      <NavBar />
+      {/* <NavBar /> */}
       <div className="container p-4">
         <Routes>
-          <Route path="/" Component={Produts}></Route>
+          <Route exact path="/" Component={LandingPage} /> 
+          <Route path="/home" Component={Home}/>
+          <Route path="/detail/:id" Component={Detail}/>
         </Routes>
       </div>
     </>
