--- client/src/Actions/index.js
@@ -1,2 +1,0 @@
-export * from "./actions";
-export * from "./actionsExport";
