+++ client/src/Components/ProductCard/ProductCard.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function ProductCard(){
+    return (
+        <div>
+            <h1>Soy una card</h1>
+            <h1>Soy una card</h1>
+        </div>
+
+    )
+}
