+++ client/src/Components/ProductsContainer/ProductsContainer.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function ProductsContainer(){
+    return (
+        <h1>Soy el contenedor de cards</h1>
+    )
+}
