--- client/src/Components/Produts/ProductsDetaild.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProductsDetaild = () => {
-  return <div>ProductsDetaild</div>;
-};
-
-export default ProductsDetaild;
