--- client/src/Components/Produts/Produts.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Produts = () => {
-  return <div>Produts</div>;
-};
-
-export default Produts;
