+++ client/src/views/Detail/Detail.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Detail(){
+    return (
+        <h1>Soy el detail</h1>
+    )
+}
