+++ client/src/views/Forms/SellProduct/SellProduct.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function SellProduct(){
+    return (
+        <h1>Soy el formulario de vender papuxs</h1>
+    )
+}
