+++ client/src/views/Home/Home.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Home(){
+    return (
+        <h1>Soy el Home</h1>
+    )
+}
