+++ client/src/views/LandingPage/LandingPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function LandingPage(){
+    return (
+        <h1>Soy la LandingPage</h1>
+    )
+}
