+++ client/src/views/ProductCategory/ProductCategory.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function ProductCategory(){
+    return (
+        <h1>Soy todos los productos de una categoria</h1>
+    )
+}
